gui material configs
The configs do not work. I want a distinct red color for
NvimTreeGitDirty config and a light green color for NvimTreeGitStaged. I
tried using the require('material').setup{ custom_highlights } but it
just does not work. On neovide, the gui I am using, colors are still
basic defaults for material deep ocean.

Not sure if this is an issue with neovide or with the way (or order) in
which my lua configs are run.