add command in comments
Instructions say to use the `set termguicolors` in order to make the
color scheme work for nvim tree. I find that this command only makes the
colors uglier. This command has no effect when I use a colorscheme for
some reason. Even though it is commented out, gonna put it here just in
case that is the source of any future issues