ref dapui file for auto open ui
the function `dap.run()` does not appear to work. Per the docs, it is
recommended to run F5 (`dap.continue()`) instead to start the debugger.
This also means that the `<leader>da` shortcut I have defined will also
not work. Just launch the debugger with F5, and close with F3, and
manually close the integrated terminal for now.